requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Image processing and embedding
torch==2.5.1
torchvision==0.20.1
sentence-transformers==3.3.1
Pillow==11.0.0
numpy==1.26.4

# Vector database
qdrant-client==1.12.1

# HTTP client
httpx==0.28.1

# Redis for caching and scheduling
redis==5.2.1
arq==0.26.1

# Async utilities
aiofiles==24.1.0
asyncio==3.4.3

# Logging and monitoring
structlog==24.4.0
prometheus-client==0.21.1

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# Development
black==24.10.0
isort==5.13.2
mypy==1.13.0
ipython==8.31.0